结束了

+++ src/main/java/cn/honorsgc/honorv2/article/ArticleController.java
@@ -8,6 +8,7 @@ import cn.honorsgc.honorv2.article.expection.*;
 import cn.honorsgc.honorv2.article.repository.ArticleCommentRepository;
 import cn.honorsgc.honorv2.article.repository.ArticleRepository;
 import cn.honorsgc.honorv2.article.repository.TagRepository;
+import cn.honorsgc.honorv2.core.CreateWish;
 import cn.honorsgc.honorv2.core.GlobalAuthority;
 import cn.honorsgc.honorv2.core.GlobalResponseEntity;
 import cn.honorsgc.honorv2.user.User;
@@ -17,18 +18,21 @@ import io.swagger.annotations.ApiParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.*;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.Authentication;
 import org.springframework.validation.Errors;
 import org.springframework.validation.FieldError;
 import org.springframework.validation.ObjectError;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
 import javax.persistence.criteria.Predicate;
-import javax.validation.Valid;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -50,30 +54,28 @@ public class ArticleController {
 
     @GetMapping({"", "/"})
     @ApiOperation(value = "查找文章")
-    public Page<ArticleSimple> index(@ApiIgnore Authentication authentication,
-                                     @ApiParam(value = "页号") @RequestParam(value = "page", required = false, defaultValue = "0") Integer pageNumber,
+    public Page<ArticleSimple> index(@ApiIgnore Authentication authentication, @ApiParam(value = "页号") @RequestParam(value = "page", required = false, defaultValue = "0") Integer pageNumber,
                                      @ApiParam(value = "类型") @RequestParam(value = "type", required = false, defaultValue = "-1") Integer type,
                                      @ApiParam(value = "用户编号") @RequestParam(value = "user", required = false, defaultValue = "-1") Integer userId,
                                      @ApiParam(value = "状态", allowableValues = "0,1,2") @RequestParam(required = false) Integer state,
                                      @ApiParam(value = "搜索文本") @RequestParam(value = "search", required = false, defaultValue = "") String search,
                                      @ApiParam(value = "使用管理员权限") @RequestParam(required = false, defaultValue = "false") Boolean admin) throws ArticleException {
         User user = (User) authentication.getPrincipal();
-
-        if (admin && state != null && user.getAuthorities().contains(GlobalAuthority.ADMIN)) {
+        admin = user.getAuthorities().contains(GlobalAuthority.ADMIN) && admin;
+        if (admin&&state != null) {
             if (state < 0 || state > 3) {
                 throw new ArticleIllegalParameterException("state参数错误");
             }
         }
 
+        Boolean finalAdmin = admin;
         Specification<Article> spec = (root, query, cb) -> {
             List<Predicate> list = new ArrayList<>();
-            logger.warn(String.valueOf(type));
             if (type >= 0) list.add(cb.equal(root.get("type"), type));
             if (userId > 0) list.add(cb.equal(root.get("user").get("id"), userId));
-
-            if (admin && state != null && user.getAuthorities().contains(GlobalAuthority.ADMIN)) {
+            if (finalAdmin&&state != null) {
                 list.add(cb.equal(root.get("state"), state));
-            } else {
+            } else if (!finalAdmin){
                 list.add(cb.or(cb.equal(root.get("state"), 1), cb.equal(root.get("user").get("id"), user.getId())));
             }
 
@@ -83,16 +85,14 @@ public class ArticleController {
         };
         Sort sort = Sort.by(Sort.Direction.DESC, "createTime");
 
-        Pageable pageable = PageRequest.of(pageNumber, 25, sort);
+        Pageable      pageable    = PageRequest.of(pageNumber, 25, sort);
         Page<Article> articlePage = articleRepository.findAll(spec, pageable);
         return articlePage.map((x) -> mapper.articleToArticleSimple(x));
     }
 
     @PostMapping({"", "/"})
     @ApiOperation(value = "编辑或者新建文章")
-    public ArticlePostResponse postArticle(@ApiIgnore Authentication authentication,
-                                           @Valid @RequestBody ArticleRequestBody requestBody,
-                                           @ApiIgnore Errors errors) throws ArticleException {
+    public ArticlePostResponse postArticle(@ApiIgnore Authentication authentication, @Validated({CreateWish.class}) @RequestBody ArticleRequestBody requestBody, @ApiIgnore Errors errors) throws ArticleException {
 
         if (errors.hasErrors()) {
             ObjectError objectError = errors.getAllErrors().get(0);
@@ -103,22 +103,11 @@ public class ArticleController {
             throw new ArticleIllegalParameterException(objectError.getDefaultMessage());
         }
 
-        User user = (User) authentication.getPrincipal();
+        User    user = (User) authentication.getPrincipal();
         Article newArticle;
-        if (requestBody.getId() != null) {
-            Optional<Article> oldArticle = articleRepository.findById(requestBody.getId());
-            if (oldArticle.isEmpty()) {
-                throw new ArticleNotFoundException();
-            }
-            if (!user.equals(oldArticle.get().getUser()) && !user.getAuthorities().contains(GlobalAuthority.ADMIN)) {
-                throw new ArticleAccessDeniedException();
-            }
-            newArticle = oldArticle.get();
-        } else {
-            newArticle = new Article();
-            newArticle.setUser(user);
-            newArticle.setCreateTime(new Date());
-        }
+        newArticle = new Article();
+        newArticle.setUser(user);
+        newArticle.setCreateTime(new Date());
         if (user.getAuthorities().contains(GlobalAuthority.ADMIN) && requestBody.getState() != null) {
             newArticle.setState(requestBody.getState());
         } else {
@@ -130,10 +119,38 @@ public class ArticleController {
         return new ArticlePostResponse("https://", savedArticle.getId());
     }
 
+    @PutMapping({"/{id}"})
+    public ArticlePostResponse updateArticle(@ApiIgnore Authentication authentication, @PathVariable Long id, @RequestBody ArticleRequestBody requestBody, @ApiIgnore Errors errors) throws ArticleException {
+        if (errors.hasErrors()) {
+            ObjectError objectError = errors.getAllErrors().get(0);
+            if (objectError instanceof FieldError) {
+                FieldError fieldError = (FieldError) objectError;
+                throw new ArticleIllegalParameterException(fieldError.getField() + fieldError.getDefaultMessage());
+            }
+            throw new ArticleIllegalParameterException(objectError.getDefaultMessage());
+        }
+        Optional<Article> oldArticle = articleRepository.findById(id);
+        if (oldArticle.isEmpty()) {
+            throw new ArticleNotFoundException();
+        }
+        User user = (User) authentication.getPrincipal();
+        if (!user.equals(oldArticle.get().getUser()) && !user.getAuthorities().contains(GlobalAuthority.ADMIN)) {
+            throw new ArticleAccessDeniedException();
+        }
+        Article article = oldArticle.get();
+        if (user.getAuthorities().contains(GlobalAuthority.ADMIN) && requestBody.getState() != null) {
+            article.setState(requestBody.getState());
+        } else {
+            article.setState(ArticleState.notApproved);
+        }
+        postArticleDo(requestBody, article);
+        Article savedArticle = articleRepository.save(article);
+        return new ArticlePostResponse("https://", savedArticle.getId());
+    }
+
     @GetMapping("/{id}")
     @ApiOperation(value = "获取文章")
-    public ArticleDto getArticle(@ApiIgnore Authentication authentication,
-                              @PathVariable Long id) throws ArticleNotFoundException {
+    public ArticleDto getArticle(@ApiIgnore Authentication authentication, @PathVariable Long id) throws ArticleNotFoundException {
         Optional<Article> article = articleRepository.findById(id);
         if (article.isEmpty()) {
             throw new ArticleNotFoundException();
@@ -147,23 +164,25 @@ public class ArticleController {
     }
 
     public void postArticleDo(ArticleRequestBody requestBody, Article article) throws ArticleIllegalParameterException {
-        Optional<Tag> tag = tagRepository.findById(requestBody.getTag());
-        if (tag.isEmpty()) {
-            throw new ArticleIllegalParameterException("tag不存在");
+        if (requestBody.getTag() != null) {
+            Optional<Tag> tag = tagRepository.findById(requestBody.getTag());
+            if (tag.isEmpty()) {
+                throw new ArticleIllegalParameterException("tag不存在");
+            }
+            article.setTag(tag.get());
         }
-        article.setType(requestBody.getType());
-        article.setTag(tag.get());
-        article.setTitle(requestBody.getTitle());
-        article.setDetail(requestBody.getDetail());
-        article.setDescribe(requestBody.getDescribe());
-        article.setHaveComment(requestBody.getHaveComment());
+
+        if (requestBody.getType() != null) article.setType(requestBody.getType());
+        if (requestBody.getTitle() != null) article.setTitle(requestBody.getTitle());
+        if (requestBody.getDetail() != null) article.setDetail(requestBody.getDetail());
+        if (requestBody.getDescribe() != null) article.setDescribe(requestBody.getDescribe());
+        if (requestBody.getHaveComment() != null) article.setHaveComment(requestBody.getHaveComment());
     }
 
     @GetMapping("/change_state")
     @Secured({"ROLE_ADMIN"})
     @ApiOperation(value = "修改文章状态")
-    public GlobalResponseEntity<String> changeArticleState(@ApiParam(value = "编号", required = true) @RequestParam Long id,
-                                                           @ApiParam(value = "状态", required = true, allowableValues = "0,1,2") @RequestParam Integer state) throws ArticleException {
+    public GlobalResponseEntity<String> changeArticleState(@ApiParam(value = "编号", required = true) @RequestParam Long id, @ApiParam(value = "状态", required = true, allowableValues = "0,1,2") @RequestParam Integer state) throws ArticleException {
         Optional<Article> article = articleRepository.findById(id);
         if (article.isEmpty()) {
             throw new ArticleIllegalParameterException("articleId 不存在");
@@ -177,9 +196,7 @@ public class ArticleController {
     //TODO: 单文章多标签
     @GetMapping("/tag")
     @ApiOperation(value = "获取文章标签")
-    public List<Tag> getTags(@RequestParam(required = false, defaultValue = "") String search,
-                             @RequestParam(required = false, defaultValue = "0") boolean admin,
-                             @ApiIgnore Authentication authentication) {
+    public List<Tag> getTags(@RequestParam(required = false, defaultValue = "") String search, @RequestParam(required = false, defaultValue = "0") boolean admin, @ApiIgnore Authentication authentication) {
         admin = admin && authentication.getAuthorities().contains(GlobalAuthority.ADMIN);
         if (admin) {
             return tagRepository.findAll();
@@ -193,24 +210,26 @@ public class ArticleController {
 
     @PostMapping("/tag")
     @ApiOperation(value = "新建标签")
-    public Tag createTag(@ApiIgnore Authentication authentication,
-                         @RequestParam String name) throws ArticleException {
+    public Tag createTag(@ApiIgnore Authentication authentication, @RequestParam String name) throws ArticleException {
         Optional<Tag> tagOptional = tagRepository.findTagByName(name);
         if (tagOptional.isPresent()) {
             throw new TagIsExistException();
         }
         User user = (User) authentication.getPrincipal();
-        Tag tag = new Tag(name, user);
+        Tag  tag  = new Tag(name, user);
         return tagRepository.save(tag);
     }
 
     @DeleteMapping("/tag/{id}")
     @Secured("ROLE_ADMIN")
     @ApiOperation(value = "删除标签")
-    public GlobalResponseEntity<String> deleteTag(@PathVariable Long id) throws TagIsNotFound {
+    public GlobalResponseEntity<String> deleteTag(@PathVariable Long id) throws ArticleException {
         Optional<Tag> optionalTag = tagRepository.findById(id);
         if (optionalTag.isEmpty()) {
             throw new TagIsNotFound();
+        }
+        if (optionalTag.get().getCount()>0){
+            throw new TagCountIsNotEmpty();
         }
         tagRepository.delete(optionalTag.get());
         GlobalResponseEntity<String> responseEntity = new GlobalResponseEntity<>();
@@ -218,27 +237,24 @@ public class ArticleController {
         return responseEntity;
     }
 
-     @GetMapping("/cmt/{id}")
+    @GetMapping("/cmt/{id}")
     @ApiOperation(value = "获取评论")
-    public ArticleCommentResponse getComment(@ApiParam(value = "文章编号", required = true) @PathVariable Long id) {
+    public List<ArticleCommentDto> getComment(@ApiParam(value = "文章编号", required = true) @PathVariable Long id) {
         List<ArticleComment> articleCommentList = articleCommentRepository.findArticleCommentsByArticle_Id(id);
-        List<ArticleCommentDto> articleCommentDtoList = mapper.articleCommentToArticleCommentDto(articleCommentList);
-        return ArticleCommentResponse.valuesOf(id, articleCommentDtoList);
+        return mapper.articleCommentToArticleCommentDto(articleCommentList);
     }
 
     @GetMapping("/cmt")
     @ApiOperation(value = "获取评论")
     @Secured({"ROLE_ADMIN"})
-    public List<ArticleCommentAdminDto> getComments(){
-        List<ArticleComment> articleComments= articleCommentRepository.findAll();
+    public List<ArticleCommentAdminDto> getComments() {
+        List<ArticleComment> articleComments = articleCommentRepository.findAll();
         return mapper.articleCommentToArticleCommentAdminDto(articleComments);
     }
 
     @PostMapping("/cmt")
     @ApiOperation(value = "发布评论")
-    public ArticleComment createComment(@ApiIgnore Authentication authentication,
-                                        @ApiParam(value = "文章编号", required = true) @RequestParam Long id,
-                                        @ApiParam(value = "评论内容", required = true) @RequestParam String detail) throws ArticleException {
+    public ArticleComment createComment(@ApiIgnore Authentication authentication, @ApiParam(value = "文章编号", required = true) @RequestParam Long id, @ApiParam(value = "评论内容", required = true) @RequestParam String detail) throws ArticleException {
         //检查文章是否存在
         Optional<Article> optionalArticle = articleRepository.findById(id);
         if (optionalArticle.isEmpty()) {
@@ -250,7 +266,7 @@ public class ArticleController {
         }
 
         ArticleComment articleComment = new ArticleComment();
-        User auth = (User) authentication.getPrincipal();
+        User           auth           = (User) authentication.getPrincipal();
         articleComment.setArticle(article);
         articleComment.setDetail(detail);
         articleComment.setUser(auth);
@@ -260,8 +276,7 @@ public class ArticleController {
 
     @DeleteMapping("/cmt")
     @ApiOperation(value = "删除评论")
-    public GlobalResponseEntity<String> deleteComment(@ApiParam(value = "评论编号") @RequestParam List<Integer> ids,
-                                                      @ApiIgnore Authentication authentication) throws ArticleException {
+    public GlobalResponseEntity<String> deleteComment(@ApiParam(value = "评论编号") @RequestParam List<Integer> ids, @ApiIgnore Authentication authentication) throws ArticleException {
 
         List<ArticleComment> articleCommentList = articleCommentRepository.findAllById(ids);
         if (articleCommentList.isEmpty()) {
@@ -284,4 +299,30 @@ public class ArticleController {
         return responseEntity;
     }
 
+    @DeleteMapping({"", "/"})
+    @ApiOperation(value = "删除文章")
+    public GlobalResponseEntity<String> deleteArticle(@ApiParam(value = "文章编号") @RequestParam List<Long> ids, @ApiIgnore Authentication authentication) throws ArticleException {
+
+        //只有管理员才能批量删除文章
+        if (ids.size() > 1 && !GlobalAuthority.isAdmin(authentication)) {
+            throw new ArticleAccessDeniedException();
+        }
+
+        //判断传参的ids大于0
+        if (ids.isEmpty()) {
+            throw new ArticleIllegalParameterException("ids不因为空");
+        }
+
+        List<Article> articles = articleRepository.findAllById(ids);
+        if (articles.isEmpty()) {
+            throw new ArticleIllegalParameterException("ids是无效的");
+        }
+
+        if (!GlobalAuthority.isAdmin(authentication) && !articles.get(0).getUser().equals(authentication.getPrincipal())) {
+            throw new ArticleAccessDeniedException();
+        }
+
+        articleRepository.deleteAll(articles);
+        return new GlobalResponseEntity<>();
+    }
 }
